% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment4.R
\docType{package}
\name{alignment4}
\alias{alignment4}
\alias{alignment4-package}
\title{Alignment 4 experiment}
\description{
Data and materials for the Alignment 4 experiment.
}
\details{
alignment4
A data package for the Alignment 4 experiment.

Use \code{data(package='alignment4')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
